requirements.txt
streamlit>=1.32.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.0.0
matplotlib>=3.7.0
reportlab>=4.0.0
opencv-python>=4.9.0.80 